// src/app/model/i-owner.ts
export interface IOwner {
  name: string
  phone: string
  email: string
  address: string
  make: string
  model: string
  year: string
  vin: string
  license_plate: string
  insurance_company: string
  policy_number: string
  expiration_date: string
  date_of_purchase: string,
  method_of_purchase: string,
  registration_type: string
}